rename hyperliquid.rs to hypercore
