+++ crates/primitives/src/explorers/mod.rs
@@ -47,8 +47,8 @@ pub mod chainflip;
 pub use chainflip::ChainflipScan;
 pub mod relay;
 pub use relay::RelayScan;
-pub mod hyperliquid;
-pub use hyperliquid::{FlowScan, HyperliquidExplorer, HypurrScan};
+pub mod hypercore;
+pub use hypercore::{FlowScan, HyperliquidExplorer, HypurrScan};
 pub mod metadata;
 pub mod stellar_expert;
 pub use metadata::{Explorer, Metadata, MultiChainExplorer};
